Add new 'active_filters_sidebar_content' view component
